Add endpoint to allocate portions of memory
The endpoint `/mem/<size_mem_in_mb>/<leak>` will allocate a portion
of memory. It receives the amount of megabytes to allocate and
optionally if it should leak memory or not (1 enables leak)

The endpoint returns some memory metrics for the process and global
system memory.

We use the gems `vmstat` and `os` to get system and process metrics
across multiple platforms. To allocate the memory, we create a
string of the given size.